# U1-M6-Summative-Markewicz-Matthew/src/main/java/com/company/RESTWebServiceGroupProject/dto/Customer.java
package com.company.RESTWebServiceGroupProject.dto;

import javax.validation.constraints.NotEmpty;
import java.util.Objects;

public class Customer {

    private int customer_id;

    @NotEmpty(message = "Please enter a valid first name.")
    private String first_name;

    @NotEmpty(message = "Please enter a valid last name.")
    private String last_name;

    @NotEmpty(message = "Please enter a valid email.")
    private String email;

    @NotEmpty(message = "Please enter a valid company.")
    private String company;

    @NotEmpty(message = "Please enter a valid phone number.")
    private String phone;

    public int getCustomer_id() {
        return customer_id;
    }

    public void setCustomer_id(Integer customer_id) {
        this.customer_id = customer_id;
    }

    public String getFirst_name() {
        return first_name;
    }

    public void setFirst_name(String first_name) {
        this.first_name = first_name;
    }

    public String getLast_name() {
        return last_name;
    }

    public void setLast_name(String last_name) {
        this.last_name = last_name;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getCompany() {
        return company;
    }

    public void setCompany(String company) {
        this.company = company;
    }

    public String getPhone() {
        return phone;
    }

    public void setPhone(String phone) {
        this.phone = phone;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Customer customer = (Customer) o;
        return customer_id == customer.customer_id &&
                Objects.equals(first_name, customer.first_name) &&
                Objects.equals(last_name, customer.last_name) &&
                Objects.equals(email, customer.email) &&
                Objects.equals(company, customer.company) &&
                Objects.equals(phone, customer.phone);
    }

    @Override
    public int hashCode() {
        return Objects.hash(customer_id, first_name, last_name, email, company, phone);
    }

    @Override
    public String toString() {
        return "Customer{" +
                "customer_id=" + customer_id +
                ", first_name='" + first_name + '\'' +
                ", last_name='" + last_name + '\'' +
                ", email='" + email + '\'' +
                ", company='" + company + '\'' +
                ", phone='" + phone + '\'' +
                '}';
    }
}