# U1-M6-Summative-Markewicz-Matthew/src/main/java/com/company/RESTWebServiceGroupProject/dto/Invoice.java
package com.company.RESTWebServiceGroupProject.dto;

import javax.validation.constraints.*;
import java.math.BigDecimal;
import java.time.LocalDate;
import java.util.Objects;

public class Invoice {

    private int invoiceId;

    @NotNull(message = "Please enter a valid customer ID.")
    private Integer customerId;

    @Past(message = "Please enter a valid order date.")
    @NotNull(message = "Please enter a valid order date.")
    private LocalDate orderDate;

    @Past(message = "Please enter a valid pick up date.")
    @NotNull(message = "Please enter a valid pick up date.")
    private LocalDate pickupDate;

    @NotNull(message = "Please enter a valid return date.")
    private LocalDate returnDate;

    @DecimalMin(value = "0.00", inclusive = true, message = "Please enter a valid late fee.")
    @Digits(integer = 7, fraction = 2, message = "Please enter a valid late fee.")
    private BigDecimal lateFee;

    public int getInvoiceId() {
        return invoiceId;
    }

    public void setInvoiceId(int invoiceId) {
        this.invoiceId = invoiceId;
    }

    public Integer getCustomerId() {
        return customerId;
    }

    public void setCustomerId(Integer customerId) {
        this.customerId = customerId;
    }

    public LocalDate getOrderDate() {
        return orderDate;
    }

    public void setOrderDate(LocalDate orderDate) {
        this.orderDate = orderDate;
    }

    public LocalDate getPickupDate() {
        return pickupDate;
    }

    public void setPickupDate(LocalDate pickupDate) {
        this.pickupDate = pickupDate;
    }

    public LocalDate getReturnDate() {
        return returnDate;
    }

    public void setReturnDate(LocalDate returnDate) {
        this.returnDate = returnDate;
    }

    public BigDecimal getLateFee() {
        return lateFee;
    }

    public void setLateFee(BigDecimal lateFee) {
        this.lateFee = lateFee;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Invoice invoice = (Invoice) o;
        return invoiceId == invoice.invoiceId &&
                Objects.equals(customerId, invoice.customerId) &&
                Objects.equals(orderDate, invoice.orderDate) &&
                Objects.equals(pickupDate, invoice.pickupDate) &&
                Objects.equals(returnDate, invoice.returnDate) &&
                Objects.equals(lateFee, invoice.lateFee);
    }

    @Override
    public int hashCode() {
        return Objects.hash(invoiceId, customerId, orderDate, pickupDate, returnDate, lateFee);
    }

    @Override
    public String toString() {
        return "Invoice{" +
                "invoiceId=" + invoiceId +
                ", customerId=" + customerId +
                ", orderDate=" + orderDate +
                ", pickupDate=" + pickupDate +
                ", returnDate=" + returnDate +
                ", lateFee=" + lateFee +
                '}';
    }
}